refactor: extract filter helpers and add tests
